requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Database
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9
greenlet>=3.0.0

# Redis
redis>=5.0.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# AWS S3 (Optional)
boto3>=1.34.34

# Email Service
sendgrid>=6.11.0

# SMS Service
twilio>=8.11.1

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
faker>=24.0.0

# Development
black>=24.3.0
flake8>=7.0.0
mypy>=1.9.0
